FIX: better fix for 32-bit support

+++ src/linux_wayland/ffi/input.rs
@@ -1,11 +1,13 @@
 // https://github.com/torvalds/linux/blob/master/include/uapi/linux/input.h
 
+use std::ffi::c_long;
+
 #[repr(C)]
 #[allow(non_camel_case_types)]
 pub struct timeval {
     // Not quite sure if these should be 32 or 64
-    pub tv_sec: i32,
-    pub tv_usec: i32,
+    pub tv_sec: c_long,
+    pub tv_usec: c_long,
 }
 
 #[repr(C)]
